\name{show-methods}
\docType{methods}
\alias{show,hypredGenome-method}
\title{Methods for Function show in Package `methods'}
\description{
  Methods for function \code{show} in Package `methods'
}
\section{Methods}{
\describe{

  \item{\code{signature(object = "hypredGenome")}}{ Show a summary of
the defined genome, that inlcudes the number and length of the
chromosomes, the number of SNP marker and QTL, and a five value summary
of the effects of the QTL (when there are QTL). }}}

\keyword{methods}

